feat: LogListItem 컴포넌트 작업 진행중
